refactor(serving): isolate cache operation deadlines
