+++ crates/skippy-server/src/frontend/local_generation.rs
@@ -1,3 +1,4 @@
+mod cache_deadline;
 mod decode_step;
 mod linear_decode;
 mod native_mtp_decode;
